add connec

+++ Genetic.cpp
@@ -6,7 +6,7 @@
 /*   By: trobicho <trobicho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/10 00:49:48 by trobicho          #+#    #+#             */
-/*   Updated: 2019/08/10 13:45:09 by trobicho         ###   ########.fr       */
+/*   Updated: 2019/10/27 04:26:38 by trobicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int		Genetic::get_generation(void)
 	return (do_get_generation());
 }
 
-int		Genetic::get_best_score(void)
+double	Genetic::get_best_score(void)
 {
 	return (do_get_best_score());
 }
