+++ People_net.cpp
@@ -6,7 +6,7 @@
 /*   By: trobicho <trobicho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/10 00:59:30 by trobicho          #+#    #+#             */
-/*   Updated: 2019/08/11 20:41:25 by trobicho         ###   ########.fr       */
+/*   Updated: 2019/10/27 03:59:35 by trobicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 using namespace std;
 
-void					People_net::set_score(int score)
+void					People_net::set_score(double score)
 {
 	do_set_score(score);
 }
 
-int						People_net::get_score(void) const
+double					People_net::get_score(void) const
 {
 	return (do_get_score());
 }
