+++ People_net.h
@@ -6,7 +6,7 @@
 /*   By: trobicho <trobicho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/08/09 15:30:03 by trobicho          #+#    #+#             */
-/*   Updated: 2019/08/11 20:41:16 by trobicho         ###   ########.fr       */
+/*   Updated: 2019/10/27 03:58:08 by trobicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ class	People_net
 {
 	public:
 		People_net(){};
-		void					set_score(int score);
-		int						get_score(void)const;
+		void					set_score(double score);
+		double					get_score(void)const;
 		int						get_answer();
 		const vector<double>&	calc_output(const vector<double> &input);
 
@@ -33,8 +33,8 @@ class	People_net
 		friend bool operator!=(const People_net &p1, const People_net &p2);
 
 	private:
-		virtual void	do_set_score(int score) = 0;
-		virtual int		do_get_score(void) const = 0;
+		virtual void	do_set_score(double score) = 0;
+		virtual double	do_get_score(void) const = 0;
 		virtual int		do_get_answer(void) = 0;
 		virtual const vector<double>&
 						do_calc_output(const vector<double> &input) = 0;
